updates to methods for getting system properties
git-svn-id: https://findbugs.googlecode.com/svn/trunk@10886 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3

+++ findbugs/src/java/edu/umd/cs/findbugs/ba/Dataflow.java
@@ -95,7 +95,7 @@ public class Dataflow <Fact, AnalysisType extends DataflowAnalysis<Fact>> {
 	}
 
 	// Maximum number of iterations before we assume there is a bug and give up.
-	private static final int MAX_ITERS = SystemProperties.getInteger("dataflow.maxiters", 97).intValue();
+	private static final int MAX_ITERS = SystemProperties.getInt("dataflow.maxiters", 97);
 
 	private String getFullyQualifiedMethodName() {
 		String methodName;
